MAT-8330: Block copying test cases if the new name would be too long (more than 250 char).
Can occur if the target measure has a duplicate named test case that is already too close to the name length limit. The dedup logic will append a new ObjectId to the new test case name, causing it to run over the limit.

Return a list of test cases that failed to copy due to new name being too long to be reported on the UI.

Signed-off-by: Joseph Kotanchik <joseph.kotanchik@semanticbits.com>